Basic data protection

+++ api/Schema/Resolvers/FriendMutationResolver.php
@@ -2,8 +2,10 @@
 
 namespace Iola\Api\Schema\Resolvers;
 
+use Iola\Api\Auth\Errors\PermissionError;
 use Iola\Api\Schema\CompositeResolver;
 use Iola\Api\Contract\Integration\FriendshipRepositoryInterface;
+use Iola\Api\Contract\Schema\ContextInterface;
 use Iola\Api\Entities\Friendship;
 use Iola\Api\Contract\Schema\Relay\EdgeFactoryInterface;
 
@@ -14,10 +16,14 @@ class FriendMutationResolver extends CompositeResolver
         EdgeFactoryInterface $edgeFactory
     ) {
         parent::__construct([
-            "addFriend" => function($root, $args) use ($friendshipRepository, $edgeFactory) {
+            "addFriend" => function($root, $args, ContextInterface $context) use ($friendshipRepository, $edgeFactory) {
                 $input = $args["input"];
                 $userId = $input["userId"]->getId();
                 $friendId = $input["friendId"]->getId();
+
+                if ($userId !== $context->getViewer()->getUserId()) {
+                    throw new PermissionError();
+                }
 
                 $friendship = $friendshipRepository->findFriendship($userId, $friendId);
                 $autoStatus = Friendship::STATUS_PENDING;
@@ -55,13 +61,17 @@ class FriendMutationResolver extends CompositeResolver
                 ];
             },
 
-            "deleteFriend" => function($root, $args) use ($friendshipRepository) {
+            "deleteFriend" => function($root, $args, ContextInterface $context) use ($friendshipRepository) {
                 $input = $args["input"];
                 $userId = $input["userId"]->getId();
                 $friendId = $input["friendId"]->getId();
-
                 $friendship = $friendshipRepository->findFriendship($userId, $friendId);
+                $friendshipOwners = [$friendship->friendId, $friendship->userId];
                 $deletedId = null;
+
+                if (!in_array($context->getViewer()->getUserId(), $friendshipOwners)) {
+                    throw new PermissionError();
+                }
 
                 if ($friendship) {
                     $friendshipRepository->deleteByIds([$friendship->id]);
