+++ api/Schema/Resolvers/UserMutationResolver.php
@@ -2,7 +2,9 @@
 
 namespace Iola\Api\Schema\Resolvers;
 
+use Iola\Api\Auth\Errors\PermissionError;
 use Iola\Api\Contract\Integration\UserRepositoryInterface;
+use Iola\Api\Contract\Schema\ContextInterface;
 use Iola\Api\Schema\CompositeResolver;
 
 class UserMutationResolver extends CompositeResolver
@@ -10,8 +12,13 @@ class UserMutationResolver extends CompositeResolver
     public function __construct(UserRepositoryInterface $userRepository)
     {
         parent::__construct([
-            "deleteUser" => function($root, $args) use ($userRepository) {
+            "deleteUser" => function($root, $args, ContextInterface $context) use ($userRepository) {
                 $userId = $args["id"]->getId();
+
+                if ($userId !== $context->getViewer()->getUserId()) {
+                    throw new PermissionError();
+                }
+
                 $userRepository->delete($userId);
 
                 return ["deletedId" => $userId];
